TelegramBOT/requirements.txt
# Smart Media AI Assistant - المتطلبات
# Updated for Python 3.12 compatibility

# Telegram Bot
pyTelegramBotAPI>=4.15.0
# python-telegram-bot is the main library, telebot is a wrapper.
python-telegram-bot>=21.0

# Media Processing
yt-dlp>=2024.04.09
ffmpeg-python>=0.2.0
moviepy>=1.0.3 # Sticking to a stable version, dev might be risky
Pillow>=10.2.0
requests>=2.31.0

# AI & LangChain (Updating to latest compatible versions)
langchain>=0.1.16
langchain-openai>=0.0.8
langchain-community>=0.0.34
openai>=1.14.2
pydantic>=2.6.4

# System & Performance (Updating numpy to a Py3.12 compatible version)
numpy>=1.26.4
psutil>=5.9.8
python-dotenv>=1.0.1

# Extra tools from original file
aiofiles>=23.2.1
pandas>=2.2.0 # Updated for Py3.12
zipfile36>=0.1.3
rarfile>=4.1
tqdm>=4.66.2
validators>=0.22.0
uvloop>=0.19.0
